refactor: use node-fetch for universal request code
