+++ lib/request/get-buffer-response-browser.js
@@ -1,0 +1,5 @@
+module.exports = getBufferResponse
+
+function getBufferResponse (response) {
+  return response.arrayBuffer()
+}
