+++ lib/request/get-buffer-response.js
@@ -1,0 +1,5 @@
+module.exports = getBufferResponse
+
+function getBufferResponse (response) {
+  return response.buffer()
+}
