+++ lib/request/http-error.js
@@ -11,7 +11,11 @@ const STATUS_CODES = {
 module.exports = class HttpError extends Error {
   constructor (message, code) {
     super(message)
-    Error.captureStackTrace(this, this.constructor)
+    // Maintains proper stack trace for where our error was thrown (only available on V8)
+    /* istanbul ignore else */
+    if (Error.captureStackTrace) {
+      Error.captureStackTrace(this, this.constructor)
+    }
     this.name = 'HttpError'
     this.code = code
     this.status = STATUS_CODES[code]
