+++ lib/request/request.js
@@ -1,135 +1,85 @@
 'use strict'
 
-module.exports = httpRequest
+module.exports = request
 
-const url = require('url')
-const http = require('http')
-const https = require('https')
-
+const fetch = require('node-fetch')
 const debug = require('debug')('octokit:rest')
 const defaults = require('lodash/defaults')
-const isStream = require('is-stream')
+const isPlainObject = require('lodash/isPlainObject')
 const pick = require('lodash/pick')
 
+const getBuffer = require('./get-buffer-response')
 const HttpError = require('./http-error')
-const isArrayBuffer = require('./is-array-buffer')
 
-function httpRequest (origRequestOptions) {
-  const requestOptions = Object.assign(
-    url.parse(origRequestOptions.url),
-    pick(origRequestOptions, 'method', 'body', 'headers', 'agent', 'timeout')
-  )
-
+function request (requestOptions) {
   debug('REQUEST:', requestOptions)
 
   // calculate content length unless body is a stream, in which case the
   // content length is already set per option
-  if (requestOptions.body && !isStream(requestOptions.body)) {
-    // stringify body unless it’s an ArrayBuffer
-    if (!isArrayBuffer(requestOptions.body) && !Buffer.isBuffer(requestOptions.body) && typeof requestOptions.body !== 'string') {
-      requestOptions.body = JSON.stringify(requestOptions.body)
-    }
-
+  if (requestOptions.body) {
     defaults(requestOptions.headers, {
-      'content-type': 'application/json; charset=utf-8',
-      'content-length': Buffer.byteLength(requestOptions.body, 'utf8')
+      'content-type': 'application/json; charset=utf-8'
     })
   }
+
+  // https://fetch.spec.whatwg.org/#methods
+  requestOptions.method = requestOptions.method.toUpperCase()
 
   // GitHub expects "content-length: 0" header for PUT/PATCH requests without body
-  if (['patch', 'put'].indexOf(requestOptions.method) >= 0 &&
-      !requestOptions.headers['content-length']) {
-    requestOptions.headers['content-length'] = 0
+  // fetch does not allow to set `content-length` header, but we can set body to an empty string
+  if (['PATCH', 'PUT'].indexOf(requestOptions.method) >= 0 && !requestOptions.body) {
+    requestOptions.body = ''
   }
 
-  if ('content-length' in requestOptions.headers) {
-    requestOptions.headers['content-length'] = parseInt(requestOptions.headers['content-length'], 10)
+  if (isPlainObject(requestOptions.body) || Array.isArray(requestOptions.body)) {
+    requestOptions.body = JSON.stringify(requestOptions.body)
   }
 
-  const reqModule = requestOptions.protocol === 'http:' ? http : https
-  delete requestOptions.protocol
+  let headers = {}
+  return fetch(requestOptions.url, pick(requestOptions, 'method', 'body', 'headers', 'timeout', 'agent'))
 
-  return new Promise((resolve, reject) => {
-    const request = reqModule.request(requestOptions, (response) => {
-      debug('STATUS: ' + response.statusCode)
-      debug('HEADERS: ' + JSON.stringify(response.headers))
+    .then(response => {
+      const contentType = response.headers.get('content-type')
 
-      const data = []
-      response.on('data', (chunk) => {
-        data.push(chunk)
-      })
-
-      /* istanbul ignore next */
-      response.on('error', (error) => {
-        reject(new HttpError(error.message, 500))
-      })
-      response.on('end', () => {
-        if (response.statusCode !== 304 && response.statusCode >= 301 && response.statusCode <= 307) {
-          origRequestOptions.url = response.headers.location
-
-          httpRequest(origRequestOptions).then(resolve, reject)
-          return
-        }
-
-        if (response.statusCode === 304 || response.statusCode >= 400 || response.statusCode < 10) {
-          reject(new HttpError(data, response.statusCode, response.headers))
-          return
-        }
-
-        const contentType = response.headers['content-type'] || ''
-
-        if (data.length === 0) {
-          return resolve({
-            meta: response.headers
-          })
-        }
-
-        if (contentType.indexOf('application/json') !== -1) {
-          return resolve({
-            meta: response.headers,
-            data: JSON.parse(Buffer.concat(data).toString())
-          })
-        }
-
-        if (/^text\/|charset=utf-8$/i.test(contentType)) {
-          return resolve({
-            meta: response.headers,
-            data: Buffer.concat(data).toString()
-          })
-        }
-
-        resolve({
-          meta: response.headers,
-          data: Buffer.concat(data)
-        })
-      })
-    })
-
-    let aborted
-    request.on('error', (error) => {
-      if (aborted) return
-      debug('REQUEST ERROR: ' + error.message)
-      reject(new HttpError(error.message, 500))
-    })
-
-    if (requestOptions.timeout) {
-      request.setTimeout(requestOptions.timeout)
-    }
-    request.on('timeout', () => {
-      debug('REQUEST ERROR: timed out')
-      request.abort()
-      aborted = true
-      reject(new HttpError('Request timeout', 504))
-    })
-
-    if (requestOptions.body) {
-      if (isStream(requestOptions.body)) {
-        return requestOptions.body.pipe(request)
+      for (const keyAndValue of response.headers.entries()) {
+        headers[keyAndValue[0]] = keyAndValue[1]
       }
 
-      request.write(Buffer.from(requestOptions.body))
-    }
+      if (response.status === 204) {
+        return
+      }
 
-    request.end()
-  })
+      if (response.status >= 400) {
+        return response.text()
+
+          .then(message => {
+            throw new HttpError(message, response.status)
+          })
+      }
+
+      if (/application\/json/.test(contentType)) {
+        return response.json()
+      }
+
+      if (!contentType || /^text\/|charset=utf-8$/.test(contentType)) {
+        return response.text()
+      }
+
+      return getBuffer(response)
+    })
+
+    .then(data => {
+      return {
+        data,
+        meta: headers
+      }
+    })
+
+    .catch(error => {
+      if (error instanceof HttpError) {
+        throw error
+      }
+
+      throw new HttpError(error.message, 500)
+    })
 }
